requirements.txt
pylint
Django
ipython
pytest-django
django-bulk-load
psycopg2
djangorestframework